Update SignupView template
